refactor(python): keep native facades within structure budget
